feat: 수정 모드 추가 #27

+++ data/src/main/java/com/project200/data/impl/ExerciseRecordRepositoryImpl.kt
@@ -15,7 +15,7 @@ class ExerciseRecordRepositoryImpl @Inject constructor(
     private val apiService: ApiService,
     @IoDispatcher private val ioDispatcher: CoroutineDispatcher
 ) : ExerciseRecordRepository {
-    override suspend fun getExerciseDetail(recordId: Int): BaseResult<ExerciseRecord> {
+    override suspend fun getExerciseDetail(recordId: Long): BaseResult<ExerciseRecord> {
         return apiCallBuilder(
             ioDispatcher = ioDispatcher,
             apiCall = { apiService.getExerciseRecordDetail(recordId) },
