+++ domain/src/main/java/com/project200/domain/repository/ExerciseRecordRepository.kt
@@ -4,5 +4,5 @@ import com.project200.domain.model.BaseResult
 import com.project200.domain.model.ExerciseRecord
 
 interface ExerciseRecordRepository {
-    suspend fun getExerciseDetail(recordId: Int): BaseResult<ExerciseRecord>
+    suspend fun getExerciseDetail(recordId: Long): BaseResult<ExerciseRecord>
 }
